back/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
uvloop==0.19.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
openai==1.98.0
redis==5.0.1
fastapi-sessions==0.3.2
pandas==2.1.4
pillow==10.1.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
psycopg2-binary==2.9.9
pytz==2023.3